refactoring jupyter insert cell code
